Cleanup multiverse
