+++ test/multiverse/suites/sidekiq/sidekiq_instrumentation_test.rb
@@ -72,6 +72,6 @@ class SidekiqInstrumentationTest < Minitest::Test
   end
 
   def test_sidekiq_ignore_retry_errors_default_is_false
-    refute NewRelic::Agent.config[:sidekiq.ignore_retry_errors], 'Expected default value for sidekiq.ignore_retry_errors to be false'
+    refute NewRelic::Agent.config[:'sidekiq.ignore_retry_errors'], 'Expected default value for sidekiq.ignore_retry_errors to be false'
   end
 end
