--- test/multiverse/suites/sidekiq_notice_only_once/Envfile
@@ -1,18 +1,0 @@
-# This file is distributed under New Relic's license terms.
-# See https://github.com/newrelic/newrelic-ruby-agent/blob/main/LICENSE for complete details.
-# frozen_string_literal: true
-
-SIDEKIQ_VERSIONS = [
-  [nil, 3.2],
-  ['7.3.9', 2.7],
-  ['6.4.0', 2.5]
-]
-
-def gem_list(sidekiq_version = nil)
-  <<-RB
-    gem 'sidekiq'#{sidekiq_version}
-    gem 'newrelic_rpm', :require => false, :path => File.expand_path('../../../../')
-  RB
-end
-
-create_gemfiles(SIDEKIQ_VERSIONS)
