--- test/multiverse/suites/sidekiq_notice_only_once/config/newrelic.yml
@@ -1,22 +1,0 @@
----
-development:
-  error_collector:
-    enabled: true
-  apdex_t: 0.5
-  monitor_mode: true
-  license_key: bootstrap_newrelic_admin_license_key_000
-  app_name: test
-  ca_bundle_path: ../../../config/test.cert.crt
-  app_name: test
-  host: localhost
-  api_host: localhost
-  port: <%= $collector && $collector.port %>
-  transaction_tracer:
-    record_sql: obfuscated
-    enabled: true
-    stack_trace_threshold: 0.5
-    transaction_threshold: 1.0
-  capture_params: false
-  log_file_path: log
-  # Enable sidekiq.ignore_retry_errors to test death_handlers registration
-sidekiq.ignore_retry_errors: true
